test(cli): split status gateway lifecycle suite
